Update Vec3 to float precision

+++ include/vec3.h
@@ -21,17 +21,17 @@ extern "C" {
 
 
 typedef struct vec3 {
-        double x, y, z;
+        float x, y, z;
 } vec3_t;
 
 
 static inline vec3_t vec3_add(vec3_t u, vec3_t v);
 static inline vec3_t vec3_sub(vec3_t u, vec3_t v);
 
-static inline vec3_t vec3_scal(vec3_t u, double t);
+static inline vec3_t vec3_scal(vec3_t u, float t);
 static inline vec3_t vec3_neg(vec3_t u);
 
-static inline double vec3_dot(vec3_t u, vec3_t v);
+static inline float vec3_dot(vec3_t u, vec3_t v);
 static inline vec3_t vec3_cross(vec3_t u, vec3_t v);
 
 static inline vec3_t vec3_cmpnt_mult(vec3_t u, vec3_t v);
